Move types into a model file
